+++ GRDB/Core/Database+Schema.swift
@@ -129,14 +129,6 @@ extension Database {
         return primaryKey
     }
     
-    /// The number of columns in the table named `tableName`.
-    ///
-    /// - throws: A DatabaseError if table does not exist.
-    @available(*, deprecated, message: "Use db.columns(in: tableName).count instead")
-    public func columnCount(in tableName: String) throws -> Int {
-        return try columns(in: tableName).count
-    }
-    
     /// The indexes on table named `tableName`; returns the empty array if the
     /// table does not exist.
     ///